refactor(harness): compose extension runtime
