+++ docs/internals/architecture/harness/coding-to-harness-migration-inventory.md
@@ -79,7 +79,7 @@ entries.
 | removed `coding.session.resource_refresh_controller` | Harness profile / Product adapter | `harness.session.SessionResourceRefreshRuntime` owns ordered prepare, reload, optional extension discovery, disabled-skill activation, bundle commit, prompt/tool rebuild, and contained refresh failure routing. Coding binds `DefaultResourceLoader`, roots, settings, current extension runner, diagnostic wording, watcher trigger, and Coding extension-runtime refresh behavior. |
 | `coding.event` | Keep product | The common runtime envelope, ordering, queue/compaction/retry/branch/metadata/package payloads, and committed-record fact live in `loushang.harness.events`. Coding retains its accepted Agent/session dictionary only as a JSON, RPC, print, TUI, extension, filtering, and render-enrichment projection. |
 | `coding.extensions.events`, `manifest`, `loader`, `contributions`, `wrapper` | Compatibility shim | Event declarations, manifest parsing, descriptor-driven loading, contribution projection, and tool wrapping live in `loushang.harness.extensions`. Coding paths preserve imports and inject Coding API/policy/legacy-event adapters. |
-| `coding.extensions.api`, `runner`, `types`, `policy`, `hooks` | Product adapter | Neutral records, registration, conflict resolution, stable route planning, observer/interceptor/reducer dispatch, resource contribution execution, binding storage/lifetimes, and generic bound/unbound runtime contexts live in Harness. Coding keeps typed model/thinking/command specialization, provider/UI callback injection, concrete permission defaults, Product result reducers, session decisions, and Agent tool-call adaptation. |
+| `coding.extensions.api`, `runner`, `types`, `policy`, `hooks` | Product adapter | Neutral records, registration, conflict resolution, stable route planning, observer/interceptor/reducer dispatch, resource contribution execution, binding storage/lifetimes, generic bound/unbound runtime contexts, and `harness.extensions.runtime.ExtensionRuntime` composition live in Harness. The runtime owns registry snapshots, commands/flags/shortcuts, generic input/event dispatch, tool wrapping, renderer/diagnostic/visibility projection, and resource discovery through injected context factories. Coding keeps descriptor loading with its API, typed model/thinking/command specialization, provider/UI callback injection, concrete permission defaults, Product result reducers, session decisions, Agent tool-call adaptation, and legacy event mapping. |
 | `coding.bootstrap` | Keep product | Product assembly. It binds Coding's persisted Product-only capability profile to Harness resource/skill activation, prompt composition, and tool/command pack ordering; it may call harness engines but should not move. |
 | `coding.runtime` | Product adapter | Generic binding leases, runtime contexts, current-session transitions, serialized operation phases, uncommitted-candidate rollback, replacement callback ordering, exclusive import staging, navigation abort scopes, and coalesced scheduling live in Harness. `harness.session.SessionLifecycleRuntime` now owns the common active-session new/restore/fork/import/replacement/disposal flow and default fork semantics. Coding keeps composition, cwd/session-file resolution and acceptance policy, its `before` user-message fork interpretation, extension event projection, diagnostics codes, transcript semantics, package operations, and index content. |
 | `coding.ui` | Never harness | Product-owned TUI adapter and screen/controller state. Shared terminal primitives belong in `loushang.tui`, not harness. |
